Remove ARG APP_PORT and the corresponding dynamic EXPOSE.
The essence of EXPOSE is to tell the person who will run the container
to use this port. Which the person may or may not use.

From the above, if literal value of the port is not know at build time
then using EXPOSE is unnecessary.

This is further buttressed by docker documentation on exposing a port:

```
functions as a type of documentation between the person who builds the image and the person who runs the container, about which ports are intended to be published. To actually publish the port when running the container, use the -p flag on docker run to publish and map one or more ports, or the -P flag to publish all exposed ports and map them to high-order ports.
```
Source: [Docker engine reference - expose](https://docs.docker.com/engine/reference/builder/#expose)